api: separate v2 api into its own class
