+++ api/v2/api_test.go
@@ -1,4 +1,4 @@
-package api
+package v2
 
 import (
 	"net/http"
