+++ api/v2/context.go
@@ -1,4 +1,4 @@
-package api
+package v2
 
 import (
 	"errors"
