+++ api/v2/context_test.go
@@ -1,4 +1,4 @@
-package api
+package v2
 
 import (
 	"net/http"
