+++ api/v2/log.go
@@ -1,4 +1,4 @@
-package api
+package v2
 
 import (
 	"github.com/gin-gonic/gin"
