+++ api/v2/routes_ipns.go
@@ -1,4 +1,4 @@
-package api
+package v2
 
 import (
 	"fmt"
