+++ api/v2/routes_payment.go
@@ -1,4 +1,4 @@
-package api
+package v2
 
 import (
 	"context"
