+++ api/v2/routes_rtfs_cluster.go
@@ -1,4 +1,4 @@
-package api
+package v2
 
 import (
 	"net/http"
